% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{aligned_mean_only_past}
\alias{aligned_mean_only_past}
\title{Aligned mean of a given vector}
\usage{
aligned_mean_only_past(y, gap)
}
\arguments{
\item{y}{response}

\item{gap}{timestamps}
}
\description{
This function does the alignment processing for a single subject,
 aligning only on observations in the past
}
