% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{alignment_response}
\alias{alignment_response}
\title{Alignment of predictions using timestamps}
\usage{
alignment_response(y, gap, id, only.past = TRUE)
}
\arguments{
\item{y}{response}

\item{gap}{offsets}

\item{id}{subject id}
}
\description{
This performs alignment of predictions using timestamps
}
