% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gest_fit.R
\name{gest_fit}
\alias{gest_fit}
\title{Fit a model to gestational data}
\usage{
gest_fit(
  x,
  y,
  tim,
  id,
  method = c("gbm", "pcurve", "intervals"),
  interval_limits = NULL,
  ...
)
}
\arguments{
\item{x}{training feature matrix}

\item{y}{training response vector}

\item{tim}{training timestamps}

\item{id}{training subject IDs, grouping variable}

\item{method}{one of "gbm", "pcurve", "intervals"}

\item{interval_limits}{if method = "intervals", a numerical vector
indicating nominal quantiles for endpoints}
}
\description{
Fit a model to gestational data as in PAPER
}
