% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gest_fit.R, R/predict.gest_fit.R
\name{predict.gest_fit}
\alias{predict.gest_fit}
\title{Make predictions for gest_fit object}
\usage{
\method{predict}{gest_fit}(fit, xnew, timnew, idnew)

\method{predict}{gest_fit}(fit, xnew, timnew, idnew)
}
\arguments{
\item{fit}{a gest_fit object}

\item{xnew}{feature matrix for prediction}

\item{timnew}{vector of timestamps}

\item{idew}{vector of subject IDs, grouping variable}
}
\description{
Make predictions for gestational data

Make predictions for gestational data
}
